pyproject.toml
[tool.poetry]
name = "машиностроение"
version = "0.1.0"
description = ""
authors = ["Raiymbek Amantayev <raiymbek.amantayev@narxoz.kz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
djoser = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
